refactor: create Game and RandomGenerator classes
